refactor(android): update package namespace from sms_to_api to smstoapi
Remove underscores from package identifiers to follow Android naming conventions.

Changes:
- Update Android namespace and applicationId in build.gradle.kts
- Update MethodChannel names in Dart code to match new package structure
- Affects service communication channel and settings storage channel

This ensures consistency across the codebase and aligns with standard Android package naming practices that discourage underscores in package names.
