+++ android/app/src/main/kotlin/com/github/dilrandi/smstoapi/SmsForwardingTask.kt
@@ -1,4 +1,4 @@
-package com.github.dilrandi.sms_to_api
+package com.github.dilrandi.smstoapi
 
 import android.content.Context
 import kotlinx.coroutines.Dispatchers
